pyproject.toml
[tool.poetry]
name = "online-platform-electronics-trading-network"
version = "0.1.0"
description = ""
authors = ["Alex19901509 <147518422+Alex19901509@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
